requirements.txt
Django
Pillow
djangorestframework
djangorestframework-simplejwt
python-decouple
drf-yasg
celery
redis
requests
psycopg2-binary